agentic-crawler-py/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-crawler"
version = "0.1.0"
description = "Agentic web crawler for FAQ and general content extraction"
requires-python = ">=3.11"
dependencies = [
  "crawlee[playwright]>=0.5",
  "anthropic>=0.40",
  "openai>=1.50",
  "pydantic>=2",
  "beautifulsoup4>=4.12",
  "lxml>=5",
  "python-dotenv>=1.0",
  "playwright>=1.45",
]

[project.optional-dependencies]
dev = ["ruff", "mypy"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
